requirements.txt
psycopg2-binary>=2.7
gunicorn>=19.9
python-telegram-bot>=12.0.0
beautifulsoup4>=4.8.0
requests>=2.22.0
